# BE_FARMX/openapi.yaml
openapi: 3.0.0
info:
  title: FarmX API
  description: API for Farm/Pond/Scan Management System
  version: 1.0.0
  contact:
    name: FarmX Team
servers:
  - url: http://localhost:4000/api
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Pond:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        farm:
          type: string
        area:
          type: number
        status:
          type: string
          enum: [active, inactive, maintenance]
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
paths:
...
  /farms/{farmId}/ponds:
    get:
      tags:
        - Ponds
      summary: Get all ponds in a farm
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: farmId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of ponds
    post:
      tags:
        - Ponds
      summary: Create pond in farm
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: farmId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: Pond 1
                area:
                  type: number
                  example: 1000
      responses:
        "201":
          description: Pond created
  /ponds/{pondId}/analytics:
    get:
      tags:
        - Analytics
      summary: Get analytics for a pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pondId
          required: true
          schema:
            type: string
        - in: query
          name: from
          schema:
            type: string
            format: date
            example: "2024-01-01"
        - in: query
          name: to
          schema:
            type: string
            format: date
            example: "2024-01-31"
      responses:
        "200":
          description: Analytics data
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  analytics:
                    type: object
                    properties:
                      period:
                        type: object
                      totalScans:
                        type: integer
                      avgMetrics:
                        type: object
                      trend:
                        type: array
                        items:
                          type: object
...
  /ponds:
    get:
      tags:
        - Ponds
      summary: Get all ponds for current user (across all farms)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: isActive
          schema:
            type: boolean
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive, maintenance]
        - in: query
          name: farmId
          schema:
            type: string
      responses:
        "200":
          description: List of ponds
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  count:
                    type: integer
                  ponds:
                    type: array
                    items:
                      $ref: "#/components/schemas/Pond"
  /ponds/{pondId}:
    get:
      tags:
        - Ponds
      summary: Get a single pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pondId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pond details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  pond:
                    $ref: "#/components/schemas/Pond"
    put:
      tags:
        - Ponds
      summary: Update a pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pondId
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                area:
                  type: number
                status:
                  type: string
                  enum: [active, inactive, maintenance]
      responses:
        "200":
          description: Pond updated
    delete:
      tags:
        - Ponds
      summary: Delete a pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pondId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pond deleted
  /ponds/{pondId}/scans:
    get:
      tags:
        - Scans
      summary: Get all scans for a pond
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: pondId
          required: true
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: List of scans
...